+++ index.html
@@ -243,7 +243,7 @@
 	</template>
 	<!-- <script src="js/functions.js"></script>
 	<script src="js/test.ts"></script> -->
-	<script src="js/main.js"></script>
+	<script src="js/main.js" type="module"></script>
 </body>
 
 </html>
